[Silver V] Title: The primes contain arbitrarily long arithmetic progressions, Time: 0 ms, Memory: 2020 KB, Score: 7 point -BaekjoonHub

+++ 백준/Silver/27907. The primes contain arbitrarily long arithmetic progressions/The primes contain arbitrarily long arithmetic progressions.cc
@@ -4,9 +4,9 @@ using namespace std;
 int main() {
   ios::sync_with_stdio(0);
   cin.tie(0);
-  long long n;
+  int n;
   cin >> n;
-  for (long long i = 0; i < n; i++)
+  for (int i = 0; i < n; i++)
     cout << 3 + 2 * i << ' ';
   return 0;
 }
